Update README.md
Added "This model was run for a maximum of 3000 iterations on the training data before being run on the testing data and compared to the ground truth labels. After testing with various powers of inverse regularization scores, the score of 1 was used to maximize run time and accuracy. The logistic fit took around ~2000 iterations to converge. If larger data sets are used in the future, the score could be brought down to e-1. This would bring the iterations to below 300 at a 1-2% cost loss in accuracy." to the implementation of log regression section